refactor(organisms): merge SfSidebar
