--- packages/vue/src/components/organisms/SfSidebar/SfSidebar.html
@@ -1,28 +1,0 @@
-<div class="sf-sidebar">
-  <SfOverlay :visible="visibleOverlay" @click="$emit('close')" />
-  <transition :name="'slide-' + position">
-    <aside v-if="visible" class="sf-sidebar__aside">
-      <div class="sf-sidebar__content">
-        <slot name="title">
-          <SfHeading
-              v-if="headingTitle"
-              :title="headingTitle"
-              :subtitle="headingSubtitle"
-              :level="headingLevel"
-              class="sf-heading--left sf-heading--no-underline sf-sidebar__title"
-          />
-        </slot>
-        <slot />
-      </div>
-      <slot name="circle-icon">
-        <SfCircleIcon
-          v-if="button"
-          icon-size="14px"
-          icon="cross"
-          class="sf-sidebar__circle-icon"
-          @click="$emit('close')"
-        />
-      </slot>
-    </aside>
-  </transition>
-</div>
